refactor: rearrange structure
